Add and refactor tests for activity and fraud questions
https://trello.com/c/reemF8gj/159-checks-on-identity-questions-in-tool
