+++ app/views/assessments/overview.html.erb
@@ -109,7 +109,7 @@
           </td>
           <td class="govuk-table__cell multi-question" scope="row">
 
-              <a href="fraud-start">Answer required</a>
+              <a href="fraud-start" id='fraud-edit'>Answer required</a>
 
           </td>
           <td class="govuk-table__cell" scope="row"></td>
