+++ spec/system/activity_questions_spec.rb
@@ -1,5 +1,5 @@
 require "rails_helper"
-require_relative "new_assessment_flow_spec"
+require_relative "steps_helper"
 
 RSpec.feature 'Activity questions', type: :system do
   scenario 'Answer activity questions positively' do
@@ -18,14 +18,6 @@ end
 
 def and_i_choose_the_activity_link
   click_link 'activity-edit'
-end
-
-def and_i_respond(response_choice, to_prompt: '', and_continue: false)
-  if to_prompt.present?
-    expect(page).to have_content to_prompt
-  end
-  choose response_choice
-  click_button 'Continue' if and_continue
 end
 
 def then_i_get_a_score(score, additional_text)
